feat: support for ACLs in grpc endpoints
