+++ akka-javasdk-maven/akka-javasdk-parent/pom.xml
@@ -7,12 +7,12 @@
     <parent>
         <groupId>io.akka</groupId>
         <artifactId>akka-javasdk-maven</artifactId>
-        <version>3.0.2</version>
+        <version>3.0.2-109-f8d7219a-dev-SNAPSHOT</version>
     </parent>
 
     <groupId>io.akka</groupId>
     <artifactId>akka-javasdk-parent</artifactId>
-    <version>3.0.2</version>
+    <version>3.0.2-109-f8d7219a-dev-SNAPSHOT</version>
     <packaging>pom</packaging>
 
 
@@ -40,7 +40,7 @@
 
         <!-- These are dependent on runtime environment and cannot be customized by users -->
         <maven.compiler.release>21</maven.compiler.release>
-        <kalix-runtime.version>1.3.1-6cd7992</kalix-runtime.version>
+        <kalix-runtime.version>1.3.1</kalix-runtime.version>
 
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <skip.docker>false</skip.docker>
